requirements.txt
autopep8==1.4.4
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
pycodestyle==2.5.0
pytz==2019.1
sqlparse==0.3.0
